+++ lib/app_metadata.dart
@@ -1,9 +1,0 @@
-import 'dart:typed_data';
-
-class AppMetadata {
-  String name;
-  String url;
-  Uint8List icon;
-
-  AppMetadata(this.name, this.url, this.icon);
-}