+++ CHANGELOG.md
@@ -1,3 +1,17 @@
-## 0.0.1
+# 1.2.0
 
-* TODO: Describe initial release.
+* **Breaking change:** `AppMetadata` is renamed to `AppHandler`.
+* Add Windows support for supported applications list.
+
+# 1.1.1
+
+* Extend `launchFile()` function.
+* Add fallback to earlier macOS version for opening files.
+
+# 1.1.0
+
+* Add 'Open with' functionality (get supported applications).
+
+# 1.0.0
+
+* Initial stable release.